net: Introduce network stack (netstack) interface
The netstack interface is a kobj(9) interface to functionality that may
be network stack specific. A kobj interface is used to allow for
supporting multiple implementations of network stacks to exist in the
future.

Additional SYSINIT levels for initializing the netstack code and for
handling registration of netstack modules.

The netstack_get_capabilities() method can be used to determine what
capabilities a netstack has. This will allow for NIC drivers and others
to determine if the network stack supports certain features.

Add a default implementation for the netstack_get_capabilities() to
advertise that no capabilities are available in the netstack.

In the FreeBSD netstack, add a netstack_get_capabilities() method that
advertises all capabilities are available.

The netstack_socreate() method is called during socreate to allow the
network stack to do any special initialization that may be needed
during socket creation.

In the FreeBSD netstack, moved setting socket FIB number from process
FIB number for PF_INET, PF_INET6 and PF_ROUTE to a netstack_socreate
method.

Obtained from:	Juniper Networks, Inc.