jail: Change prison address handling to be AF-neutral
Instead of only being enabled with INET or INET6 is enabled, allow for
jail address handling to be enabled always.
Change the naming of prison_ip_* functions to prison_addr_* instead.
Change PR_IP() macro name to PR_ADDR()

This allows for other address families to be supported in the future or
downstream vendor implementations.

Obtained from:	Juniper Networks, Inc.